Move kubeval task, add merge file to pipeline-agent, add kubeval test
