+++ envs/minikube/merges/service/pipeline-agent/kube/01-config.yaml
@@ -1,0 +1,5 @@
+# THIS FILE HAS BEEN AUTOMATICALLY GENERATED BY: poetry run invoke update-from-templates
+# GENERATED FROM service/pipeline-agent/kube/merge-templates/01-config.yaml, DO NOT MODIFY THIS FILE BY HAND
+#
+data:
+  ENV: "{env}"
