+++ envs/minikube/settings.py
@@ -3,4 +3,4 @@ COMPONENTS = ["service/pipeline-agent"]
 KUBE_CONTEXT = "minikube"
 KUBE_NAMESPACE = "default"
 IMAGE_PULL_SECRETS = {}
-TEMPLATE_VARIABLES = {}
+TEMPLATE_VARIABLES = {"env": "development"}
