+++ service/pipeline-agent/kube/merge-templates/01-config.yaml
@@ -1,0 +1,2 @@
+data:
+  ENV: "{env}"
