+++ tasks.py
@@ -259,27 +259,7 @@ def kubeval(ctx):
     Check that all Kubernetes configs look valid with kubeval
     :param Context ctx:
     """
-
-    label(logger.info, "Checking Kubernetes configs")
-
-    def _should_ignore(path):
-        parts = path.parts
-        if parts[0] == "temp":
-            return True
-        elif parts[0] == "envs" and parts[2] == "merges":
-            return True
-
-        return False
-
-    kube_yamls = [
-        str(path)
-        for path in Path(".").glob("**/kube/*.yaml")
-        if not _should_ignore(path)
-    ]
-
-    skip_kinds = ",".join(devops.settings.KUBEVAL_SKIP_KINDS)
-
-    run(["kubeval", "--skip-kinds", skip_kinds] + kube_yamls)
+    devops.tasks.kubeval()
 
 
 @task()
